some fixes. add tests

+++ src/internal/server/http/v1/requests/bid/get_list.go
@@ -5,8 +5,8 @@ import "github.com/google/uuid"
 type GetListRequest struct {
 	TenderID uuid.UUID `json:"tenderId" validate:"required,uuid,max=100"`
 	Username string    `query:"username" validate:"required,max=50"`
-	Limit    uint      `query:"limit" validate:"gte=0"`
-	Offset   uint      `query:"offset" validate:"gte=0"`
+	Limit    int64     `query:"limit" validate:"gte=0"`
+	Offset   int64     `query:"offset" validate:"gte=0"`
 }
 
 func (r *GetListRequest) SetDefaults() {
