+++ src/internal/server/http/v1/requests/bid/get_my.go
@@ -2,8 +2,8 @@ package bid
 
 type GetMyRequest struct {
 	Username string `query:"username" validate:"required,max=50"`
-	Limit    uint   `query:"limit" validate:"gte=0"`
-	Offset   uint   `query:"offset" validate:"gte=0"`
+	Limit    int64  `query:"limit" validate:"gte=0"`
+	Offset   int64  `query:"offset" validate:"gte=0"`
 }
 
 func (r *GetMyRequest) SetDefaults() {
