+++ src/internal/server/http/v1/requests/bid/get_reviews.go
@@ -6,8 +6,8 @@ type GetReviewsRequest struct {
 	TenderID          uuid.UUID `json:"tenderId" validate:"required,uuid,max=100"`
 	AuthorUsername    string    `query:"authorUsername" validate:"required,max=50"`
 	RequesterUsername string    `query:"requesterUsername" validate:"required,max=50"`
-	Limit             uint      `query:"limit" validate:"gte=0"`
-	Offset            uint      `query:"offset" validate:"gte=0"`
+	Limit             int64     `query:"limit" validate:"gte=0"`
+	Offset            int64     `query:"offset" validate:"gte=0"`
 }
 
 func (r *GetReviewsRequest) SetDefaults() {
