+++ src/internal/server/http/v1/requests/tender/get_my.go
@@ -1,8 +1,8 @@
 package tender
 
 type GetMyRequest struct {
-	Limit    uint   `query:"limit" validate:"gte=0"`
-	Offset   uint   `query:"offset" validate:"gte=0"`
+	Limit    int64  `query:"limit" validate:"gte=0"`
+	Offset   int64  `query:"offset" validate:"gte=0"`
 	Username string `query:"username" validate:"required,max=50"`
 }
